Move code inside module
